requirements.txt
streamlit
simplekml
pyproj
Pillow
pytesseract
numpy